+++ src/graphql/client/derives.rs
@@ -1,11 +1,13 @@
 #![allow(clippy::enum_variant_names)]
 
-use chrono::{DateTime as ChronoDateTime, Utc};
 use graphql_client::GraphQLQuery;
 
-type DateTime = ChronoDateTime<Utc>;
+use crate::graphql::{
+    StringNumberI64, StringNumberU32, StringNumberU64, StringNumberUsize, TimestampIso8601,
+};
 
-use crate::graphql::{StringNumberI64, StringNumberU32, StringNumberU64, StringNumberUsize};
+// Type alias for graphql_client generated code
+type Timestamp = TimestampIso8601;
 
 #[derive(GraphQLQuery)]
 #[graphql(